[PI-40][chore] Split scaffold tests and unify agent instructions
