+++ .github/copilot-instructions.md
@@ -1,6 +1,6 @@
 # GitHub Copilot Instructions
 
-Full agent rules are in [AGENTS.md](../AGENTS.md) and repo conventions in [CLAUDE.md](../CLAUDE.md).
+Full agent rules and repo conventions are in [CLAUDE.md](../CLAUDE.md). [AGENTS.md](../AGENTS.md) redirects there.
 
 ## Quick reference (Copilot Workspace / inline chat)
 
@@ -8,7 +8,8 @@ Full agent rules are in [AGENTS.md](../AGENTS.md) and repo conventions in [CLAUD
 - Tracking system: **GitHub Projects** (board) + **GitHub Issues** (tickets)
 - Create issues: `gh issue create` — pick the right template (bug / feature / chore)
 - Board cards move automatically via `board-automation.yml` — no manual updates needed
-- Issue, branch, and PR names use the Project Init key: `PI-<issue-number>`, e.g. `PI-42`
+- Branch names use `<issue_type>/<project_abbr>-<issue_number>-<branch-short-description>`, e.g. `chore/PI-40-split-scaffold-tests`
+- Issue and PR names use the Project Init key: `PI-<issue-number>`, e.g. `PI-42`
 - PR titles must follow: `[PI-N][type] description` where type ∈ {feat, fix, chore, docs, test}, e.g. `[PI-42][feat] Add OAuth login`
 - For small no-issue PRs: `[nojira][type] description`, e.g. `[nojira][fix] Fix typo`
 - PR body must still include the GitHub numeric reference `Closes #N` — auto-closes issue and moves board card to Done on merge (skip for nojira PRs)
@@ -25,4 +26,4 @@ Full agent rules are in [AGENTS.md](../AGENTS.md) and repo conventions in [CLAUD
 - Core dep is `rich`; stdlib (`tomllib`, `argparse`, `pathlib`) covers the rest — don't add deps without justification
 - Templates live under `templates/` using `dot_claude/` / `dot_gitignore` naming; scaffolder renames on copy
 - Template placeholders use `{{variable_name}}` syntax
-- Any change to `templates/` needs a corresponding test that scaffolds into a temp dir and diffs output
+- Any change to `templates/` needs a corresponding test that scaffolds into a temp dir and diffs output. Put new tests in the focused `tests/test_*.py` file for that behavior; create a new focused test file if needed.
