+++ CLAUDE.md
@@ -2,7 +2,9 @@
 
 This repo is a **scaffolder**. It produces a `.claude/` layout inside *other* projects. Nothing here runs as a long-lived service.
 
-Start by reading [AGENTS.md](AGENTS.md), then use [README.md](README.md) for user-facing behavior and [.github/copilot-instructions.md](.github/copilot-instructions.md) for GitHub Issues, PR titles, PR bodies, and board behavior.
+This is the canonical instruction file for agents working in this repository. [AGENTS.md](AGENTS.md) and [GEMINI.md](GEMINI.md) intentionally redirect here to avoid duplicated rules.
+
+Use [README.md](README.md) for user-facing behavior and [.github/copilot-instructions.md](.github/copilot-instructions.md) for GitHub Issues, PR titles, PR bodies, and board behavior.
 
 ## Layout
 
@@ -15,7 +17,7 @@ Start by reading [AGENTS.md](AGENTS.md), then use [README.md](README.md) for use
 │   ├── obsidian/           # overlay for both Obsidian-* presets
 │   ├── lightrag/           # overlay for Obsidian+LightRAG preset
 │   └── presets/            # toml preset definitions
-└── tests/
+└── tests/                  # focused pytest modules by behavior area
 ```
 
 Template naming convention: directories stored as `dot_claude/`, `dot_gitignore` etc. The scaffolder renames them to `.claude/`, `.gitignore` on copy. This keeps templates visible in GitHub and avoids this repo being auto-loaded as a Claude Code config for itself.
@@ -26,7 +28,15 @@ Template naming convention: directories stored as `dot_claude/`, `dot_gitignore`
 - **Deterministic** — copy/render logic is pure file ops; never call an LLM from the scaffolder itself.
 - **uv everywhere** — `uv run …`, never `pip install` or `python -m venv`.
 - **ruff only** — no black / isort / mypy.
-- **Templates are tested by scaffolding into a temp dir** — any change to `templates/` should have a corresponding test that runs the wizard with a preset and diffs the output.
+- **Templates are tested by scaffolding into a temp dir** — any change to `templates/` should have a corresponding test in the focused `tests/test_*.py` module for that behavior. Create a new focused file if no existing module fits.
+
+## GitHub workflow
+
+- Work is tracked in GitHub Projects backed by GitHub Issues.
+- Branch names must follow `<issue_type>/<project_abbr>-<issue_number>-<branch-short-description>`, for example `chore/PI-40-split-scaffold-tests`.
+- Project Init uses `PI` as the project abbreviation.
+- PR titles must follow `[PI-IssueNumber][type] description` where type is one of `feat`, `fix`, `chore`, `docs`, or `test`.
+- PR bodies must include `Closes #<number>` to auto-link the issue and move the board card on merge.
 
 ## CI Optimizations
