+++ templates/base/AGENTS.md.tmpl
@@ -1,66 +1,5 @@
-# AGENTS.md — layout spec for AI coding agents
+# AGENTS.md
 
-This project was scaffolded by [project-init]({{project_init_url}}) and follows a model-agnostic layout so that any coding agent (Claude Code, Cursor, Aider, Codex, Gemini, etc.) can navigate it.
+Canonical instructions for this project live in [CLAUDE.md](CLAUDE.md).
 
-## Where things live
-
-```
-.
-├── CLAUDE.md             # Claude Code entry point (thin — redirects here)
-├── AGENTS.md             # this file
-└── .claude/
-    ├── project-init.md   # project workflow and conventions
-    ├── config.yaml       # tools/MCPs selected at init
-    ├── settings.json     # Claude Code settings (safe for other agents to ignore)
-    ├── memory/
-    │   ├── MEMORY.md     # index of memory files — READ FIRST for context
-    │   └── *.md          # small structured memory files (user/feedback/project/reference)
-    ├── vault/            # Obsidian vault — human-authored docs
-    │   ├── decisions/    # architectural decisions (ADRs)
-    │   ├── design/       # design notes
-    │   ├── sessions/     # session logs
-    │   └── knowledge/    # domain notes
-    ├── commands/         # Claude Code slash commands (markdown)
-    ├── skills/           # Claude Code skills (markdown)
-    ├── agents/           # Claude Code subagents (markdown)
-    ├── hooks/            # deterministic session hooks (bash/python)
-    └── scripts/          # helper scripts (bash/python)
-```
-
-## Orientation protocol (all agents)
-
-When starting work on this project:
-
-1. Read [`.claude/project-init.md`](.claude/project-init.md) for the workflow and conventions.
-2. Read [`.claude/memory/MEMORY.md`](.claude/memory/MEMORY.md) to load prior context.
-3. Browse [`.claude/vault/`](.claude/vault/) for design/decision notes relevant to your task.
-4. Check [`.claude/config.yaml`](.claude/config.yaml) to see which MCPs/tools are installed.
-
-## Memory conventions
-
-Memory files in `.claude/memory/` use a simple frontmatter schema:
-
-```markdown
----
-name: <short title>
-description: <one-line summary used for relevance ranking>
-type: user | feedback | project | reference
----
-
-<body>
-```
-
-`MEMORY.md` is a plain index — one bullet per memory file, under ~150 chars per line.
-
-See [`.claude/memory/README.md`](.claude/memory/README.md) for the full convention.
-
-## Per-agent notes
-
-- **Claude Code**: `.claude/` is native. `CLAUDE.md` at root is auto-loaded.
-- **Cursor**: create `.cursorrules` with `"Read AGENTS.md and .claude/project-init.md before starting any task."`
-- **Aider**: add to `.aider.conf.yml` → `read: [AGENTS.md, .claude/project-init.md]`
-- **Codex / other**: point the agent at `AGENTS.md` first.
-
-## Deterministic-first principle
-
-Hooks and scripts in `.claude/` prefer bash/python over LLM calls. Reserve LLM calls for genuinely generative tasks (summarization, entity extraction, judgment). Everything else — file ops, pattern matching, template rendering, graph traversal — is deterministic code.
+Read [CLAUDE.md](CLAUDE.md) before working in this codebase. It is the source of truth for workflow, conventions, memory, tools, and branch naming.
