+++ templates/base/CLAUDE.md.tmpl
@@ -10,7 +10,7 @@ All agentic-development infrastructure lives under [`.claude/`](.claude/). Start
 - [`.claude/memory/MEMORY.md`](.claude/memory/MEMORY.md) — memory index (read first for context)
 - [`.claude/docs/`](.claude/docs/) — system of record: ADRs and development guides
 - [`.claude/vault/`](.claude/vault/) — human workspace (exploratory notes)
-- [`AGENTS.md`](AGENTS.md) — layout spec for non-Claude agents (Cursor, Aider, Codex, etc.)
+- [`AGENTS.md`](AGENTS.md) and [`GEMINI.md`](GEMINI.md) — thin redirects here for non-Claude agents
 
 > Scaffolded with [project-init]({{project_init_url}}) on {{created_date}}.
 
@@ -18,6 +18,7 @@ All agentic-development infrastructure lives under [`.claude/`](.claude/). Start
 
 - **TDD** — write failing tests before any implementation. Use `/plan <task>` to generate acceptance tests first.
 - **GitHub Projects** — work is tracked on the GitHub Projects board (kanban) backed by GitHub Issues. Before starting non-trivial work, create or reference an issue. Use `/start-task` if one does not exist.
+- **Branch names** — use `<issue_type>/<project_abbr>-<issue_number>-<branch-short-description>`, for example `feat/PI-42-add-oauth-login`.
 {{#if lint_command}}- **Lint** — `{{lint_command}}` must pass before closing a task. The `pre-commit-gate` hook enforces this on every commit.
 {{/if}}- **No secrets in code** — never hardcode API keys, tokens, or personal data. Use `os.environ['KEY']` or a `.env` file (gitignored).
