+++ templates/base/dot_claude/config.yaml.tmpl
@@ -6,7 +6,7 @@ project:
   description: "{{project_description}}"
   created: {{created_date}}
   project_init_version: {{project_init_version}}
-  project_key: ""   # e.g. "PI" → commit/PR titles become [PI-42][feat]; blank uses [#42]
+  project_key: ""   # e.g. "PI" -> branches use feat/PI-42-slug and PR titles use [PI-42][feat]; blank derives from repo name
 
 language: {{language}}            # python | node | go | none
