+++ templates/base/dot_claude/skills/start-task/SKILL.md
@@ -24,7 +24,7 @@ Before starting any non-trivial task, create a GitHub Issue, a dedicated branch,
    ```bash
    .claude/scripts/start-issue.sh <issue-number> <type>
    ```
-   This derives the branch name (`<type>/<n>-<slug>`) from the issue title, pushes it, opens a draft PR with the correct `[#n][type]` title and `Closes #n` body, and arms auto-merge if the repo supports it.
+   This derives the branch name (`<issue_type>/<project_abbr>-<issue_number>-<slug>`) from the issue title, pushes it, opens a draft PR with the correct `[PROJECT-123][type]` title and `Closes #n` body, and arms auto-merge if the repo supports it.
 
 5. **Proceed** — only begin implementation after the scripts have run successfully.
