+++ templates/base/dot_github/hooks/pre-push
@@ -10,7 +10,7 @@ while read local_ref local_sha remote_ref remote_sha; do
     echo "❌ ERROR: Direct push to main/master is not allowed"
     echo ""
     echo "To contribute:"
-    echo "  1. Create a new branch with the project key: git checkout -b <PROJECT-KEY>-<issue-number>-<slug>"
+    echo "  1. Create a new branch with the project key: git checkout -b <issue_type>/<project_abbr>-<issue_number>-<slug>"
     echo "  2. Make your changes and push: git push origin <branch-name>"
     echo "  3. Open a PR on GitHub"
     echo ""
