+++ tests/conftest.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture
+def tmp_target(tmp_path: Path) -> Path:
+    return tmp_path / "project"
