+++ tests/test_presets.py
@@ -1,0 +1,30 @@
+from __future__ import annotations
+
+import pytest
+
+from project_init.scaffold import list_presets, load_preset
+
+
+class TestListPresets:
+    def test_returns_both_presets(self):
+        presets = list_presets()
+        names = {p["name"] for p in presets}
+        assert "obsidian-only" in names
+        assert "obsidian-lightrag" in names
+
+    def test_presets_have_required_keys(self):
+        for p in list_presets():
+            assert "name" in p
+            assert "description" in p
+            assert "layers" in p
+
+
+class TestLoadPreset:
+    def test_load_known_preset(self):
+        p = load_preset("obsidian-only")
+        assert p["name"] == "obsidian-only"
+        assert "base" in p["layers"]
+
+    def test_load_unknown_preset_raises(self):
+        with pytest.raises(ValueError, match="Unknown preset"):
+            load_preset("nonexistent")
